Fix bugs in Alfred's install script

+++ alfred/install.sh
@@ -3,21 +3,32 @@ set -e
 
 mkdir -p ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences
 
-dirs=(appearence features)
+dirs=(appearance features)
 
 for dir in "${dirs[@]}"
 do
     rm -rf ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/$dir || true
-    ln -s $ZSH/alfred/appearence ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/$dir
+    ln -s $ZSH/alfred/$dir ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/$dir
 done
+
+if [ -d ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/local ]
+then
+
+fi
+
+if [[ ! -d ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/local ]]
+then 
+    echo "Save at least 1 setting for Alfred to bootstrap pref dir, then re-run $ZSH/alfred/install.sh"
+    exit 1
+fi
 
 localdir=$(/bin/ls -1 ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/local)
 
 rm -rf ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/local/$localdir/* || true
 
-dirs=(appearence features hotkey)
+dirs=(appearance features hotkey)
 
 for dir in "${dirs[@]}"
 do
-    ln -s $ZSH/alfred/local/appearence ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/local/$localdir/$dir
+    ln -s $ZSH/alfred/local/$dir ~/Library/Application\ Support/Alfred/Alfred.alfredpreferences/preferences/local/$localdir/$dir
 done
